pyproject.toml
[project]
name = "kg-langgraph-agent"
version = "0.1.0"
description = "LangGraph agent scaffold with local tools, MCP tools, and skill loading."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.3.0",
  "langgraph>=0.3.0",
  "langgraph-checkpoint-sqlite>=2.0.0",
  "pyyaml>=6.0.1",
  "uvicorn>=0.30.0",
  "arq>=0.26.0",
  "docker>=7.1.0",
]

[project.optional-dependencies]
mcp = [
  "langchain-mcp-adapters>=0.1.0",
]

[project.scripts]
kg-agent = "kg_agent.cli:main"
kg-agent-web = "kg_agent.webapp:main"
kg-agent-worker = "kg_agent.worker:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
